Move CustomStatsList out into separate components
