+++ src/containers/StatOptions/types.ts
@@ -1,0 +1,4 @@
+export type StatOption = {
+  statId: string;
+  enabled: boolean;
+};
